methods/deduplication/presto-0.5.2/requirements.txt
numpy>=1.8
scipy>=0.14
pandas>=0.15
biopython>=1.65
